size_t revenge

+++ include/Events/UserEvents.hpp
@@ -27,7 +27,7 @@ namespace CoffeeMaker {
 
     private:
     unsigned int _id;
-    unsigned int _index;
+    size_t _index;
     static unsigned int _uid;
     static std::vector<IUserEventListener*> _listeners;
     static void ProcessUserEvent(const SDL_UserEvent& event);
